FRONTEND UPDATES AND INV,SALES

+++ database/migrations/2024_01_01_000013_create_inventories_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
             $table->id();
             $table->unsignedBigInteger('product_id');
             $table->unsignedBigInteger('branch_id');
-            $table->integer('available_stock')->nullable();
+            $table->decimal('available_stock')->nullable();
             $table->decimal('cost', 12, 2)->nullable();
             $table->integer('reorder_level')->nullable();
             $table->timestamps();
